% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.multiplicative.LC.multi.R
\name{forecast.multiplicative.LC.multi}
\alias{forecast.multiplicative.LC.multi}
\title{FUNCTION TO FORECAST MULTIPLICATIVE MULTI-POPULATION MORTALITY MODEL}
\usage{
forecast.multiplicative.LC.multi(
  fitted.obj,
  nahead,
  ktmethod = c("Arimapdq", "arima010"),
  kt_include.cte = TRUE
)
}
\arguments{
\item{fitted.obj}{object developed using function fit.mutliplicative.LC.multi()}

\item{nahead}{number of periods ahead to forecast}

\item{ktmethod}{method used to forecast the value of kt Arima(p,d,q) or ARIMA(0,1,0); c("Arimapdq", "arima010").}

\item{kt_include.cte}{if you want that kt include constant in the arima process.}
}
\value{
A list with different components of the forecasting process:
\item{ax}{ parameter that captures the average shape of the mortality curve in all considered populations.}

\item{bx}{ parameter that explains the age effect x with respect to the general trend (kt) in the mortality rates of all considered populations.}

\item{arimakt}{ the arima selected for the \eqn{k_t} time series.}

\item{kt.fitted}{ obtained \eqn{k_t} values for the tendency behaviour.}

\item{kt.fut}{ \eqn{k_t} projected values for the nahead periods ahead.}

\item{kt.futintervals}{ \eqn{k_t} arima selected and future values for this parameter with the different intervals, lower and upper, 80\\% and 90\\%.}

\item{Ii}{ gives an idea of the differences in the pattern of mortality in any region i with respect to Region 1.}

\item{formula}{ additive multi-population mortality formula used to fit the mortality rates.}

\item{qxt.real}{ real mortality rates.}

\item{qxt.fitted}{ fitted mortality rates using the additive multi-population mortality model.}

\item{logit.qxt.fitted}{ fitted mortality rates in logit way estimated with the additive multi-population mortality model.}

\item{qxt.future}{ future mortality rates estimated with the additive multi-population mortality model.}

\item{logit.qxt.future}{ future mortality rates in logit way estimated with the additive multi-population mortality model.}

\item{nPop}{ provided number of populations to fit the periods.}
}
\description{
R function for forecasting multiplicative multi-population mortality model developed by: Russolillo et al (2011).
This model follows the structure of the well-known Lee-Carter model (Lee and Carter, 1992) but including a multiplicative parameter to capture the behavior of each population considered.
This parameter seeks to capture the individual behavior of every population considered.
It should be mentioned that in case that this function is developed for forecasting several populations.
However, in case you only consider one population, the function will forecast the Lee-Carter model for one population.
}
\examples{
#First, we present the data that we are going to use
SpainRegions
ages <- c(0, 1, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60, 65, 70, 75, 80, 85, 90)

#Before forecast the future value of the mortality rates for different populations
#we need the object the fitted values of the multiplicative multi-population mortality model.
multiplicative_Spainmales <- fit.multiplicative.LC.multi(qxt = SpainRegions$qx_male,
                                             periods = c(1991:2020),
                                             ages = c(ages),
                                             nPop = 18,
                                             lxt = SpainRegions$lx_male)
#Once, we have fit the data, it is possible to see the \eqn{\alpha_x}, \eqn{\beta_x}, \eqn{k_t}, and \eqn{I_i} provided parameters for the fitting.
plot.fit.LC.multi(multiplicative_Spainmales)

#Finally, we forecast 10 years ahead the multiplicative multi-population mortality model

fut_multiplicative_Spainmales <- forecast.multiplicative.LC.multi(fitted.obj = additive_Spainmales, nahead = 10,
                                                      ktmethod = "Arimapdq", kt_include.cte = TRUE)
#As we mentioned in the details of the function, if we only provide the data from one-population the function
#\code{\link{fit.multiplicative.LC.multi}} will fit the Lee-Carter model for single populations.
LC_Spainmales <- fit.multiplicative.LC.multi(qxt = SpainNat$qx_male,
                              periods = c(1991:2020),
                              ages = ages,
                              nPop = 1)
plot.fit.multiplicative.LC.multi(LC_Spainmales)
#Again, we can forecast 10 years ahead using the LC mortality model for one-single population.
fut_LC_Spainmales <- forecast.multiplicative.LC.multi(fitted.obj = LC_Spainmales,
        nahead = 10,ktmethod = "Arimapdq", kt_include.cte = TRUE)

}
\references{
Lee, R.D. & Carter, L.R. (1992).
Modeling and forecasting US mortality.
Journal of the American Statistical Association, 87(419), 659–671.

Russolillo, M., Giordano, G., & Haberman, S. (2011).
Extending the Lee–Carter model: a three-way decomposition.
Scandinavian Actuarial Journal, 2011(2), 96-117.
}
\seealso{
\code{\link{fit.additive.LC.multi}}, \code{\link{fit.multiplicative.LC.multi}},
\code{\link{forecast.additive.LC.multi}}, \code{\link{multipopulation_cv}},
}
