% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpainMap.R
\name{SpainMap}
\alias{SpainMap}
\title{Spain National map information}
\usage{
SpainMap
}
\arguments{
\item{regionvalue}{vector with the values that you want to plot in percentiles in the Spain map.}

\item{main}{the specific title of the map plot}

\item{name}{the assigned name for the legend in map plot.}
}
\value{
a map from the regions of Spain colored with the variable provided by the user.
}
\description{
This data contains information to plot the panish regions using
}
\examples{
ages <- c(0, 1, 5, 10, 15, 20, 25, 30, 35, 40,
45, 50, 55, 60, 65, 70, 75, 80, 85, 90)
#We fit the additive multi-population mortality model for spain males
additive_Spainmales <- fit.additive.LC.multi(qxt = SpainRegions$qx_male,
                              periods = c(1991:2020),
                              ages = c(ages),
                              nPop = 18,
                              lxt = SpainRegions$lx_male)
name <- c("Ii")
main <- c("Multiplicative for males")
regionvalue <- additive_Spainmales$Ii[2:18]

SpainMap(regionvalue, main, name)

}
\references{
Spanish National Institute of Statistics (INE) (2023). Tablas de mortalidad, metodologia.
Technical report, Instituto Nacional de Estadistica
}
